Add pagination and search for BookInstance
Use aggregate pipeline for the mongoose query to search a book instance, by name. Book is a nested field in BookInstance. Also changed the search form so that get request on that triggers get request to url passed down to page rather than the page it's currently at (useful if we search for book name in say 2nd or 3rd page and after search, the results fit in 1 page only